Move player::disp_info and print_encumbrance
